Refactor login and registration tests: enhance docstrings and reorganize user data
