+++ ecomus/test/test_register_login.py
@@ -1,16 +1,7 @@
 import pytest
 from playwright.async_api import Page
 from models.register_login_users import RegisterLoginUser
-from utils.users_data import register_login_valid_new_users, register_login_invalid_new_users
-
-"""_summary_
-The following tests were performed:
-    1. Registration and login success: Verifies that registration is successful 
-        and redirects to the login page, followed by a successful login.
-    2. Registration and login with empty or invalid fields: Verifies that
-        registration fails if there are empty or invalid fields, and that login
-        is not successful with these credentials.
-"""
+from utils.register_login_data import register_login_valid_new_users, register_login_invalid_new_users
 
 @pytest.mark.asyncio
 @pytest.mark.parametrize(
@@ -19,6 +10,7 @@ The following tests were performed:
     indirect=["auto_delete_user"]
 )
 async def test_successful_registration_login_valid(register_page: Page, login_page: Page, user: RegisterLoginUser, auto_delete_user):
+    """Test for successful registration and login."""
     await register_page.register(user.register)
     assert "/login" in register_page.page.url, f"Valid user not created: {user.register}."
     await login_page.login(user.login)
@@ -31,6 +23,7 @@ async def test_successful_registration_login_valid(register_page: Page, login_pa
     indirect=["auto_delete_user"]
 )
 async def test_successful_registration_login_invalid(register_page: Page, login_page: Page, user: RegisterLoginUser, auto_delete_user):
+    """Test for registration and login with invalid credentials."""
     await register_page.register(user.register)
     await login_page.login(user.login)
     assert "/my-account" in login_page.page.url, f"You logged in with invalid credentials: {user.register}."
